Use defaults from external/sqlite
b/163598610 was blocking the bindgen step from referencing the cflags
from external/sqlite. The bug has now been fixed, so we can simply
reference the defaults from the library for which we're generating
rust bindings.

cargo2android.py also updated TEST_MAPPING, so I've included that in
the patch as well.

Bug: 172013262
Test: atest system/security/keystore2
Change-Id: I8d73a9e0228ed76d56b0118535f30d5f20432945